% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{acceleration_fun}
\alias{acceleration_fun}
\title{acceleration_fun}
\usage{
acceleration_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\description{
accelerated solar PV "ideation" mutiplier for agents. The 2010-2022 calibrated rate p. is replaced p. * acceleration_fun to allow for the impact of advertisinga campaigns etc.
where acceleration = 1 for yeartime < 2023.
}
