% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{amort}
\alias{amort}
\title{amort}
\usage{
amort(r, term)
}
\arguments{
\item{r}{interest rate, decimal}

\item{term}{term of loan or investment}
}
\value{
real value
}
\description{
amortisation or annuity payment for unit principal
}
