% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{amort_c}
\alias{amort_c}
\title{amort}
\usage{
amort_c(r, term)
}
\arguments{
\item{r}{decimal interest rate}

\item{term}{term of loan or investment}
}
\value{
real value
}
\description{
amortisation or annuity payment for unit principal
}
