% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills_oo.R
\docType{data}
\name{bills_oo}
\alias{bills_oo}
\title{raw bi-monthly bill information from 2018 survey}
\format{
vector with 759 elements:
\describe{
\item{q9_1}{bill in euros}
}
}
\source{
2018 survey
}
\usage{
bills_oo
}
\description{
raw bi-monthly bill information from 2018 survey
}
\keyword{datasets}
