% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bungalow_probability.R
\docType{data}
\name{bungalow_probability}
\alias{bungalow_probability}
\title{bungalow_probability}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{region}{One of Dublin, Leinster, Munster, Connulster}
\item{house_type}{One of Detached house, Semi-detached house, Terraced house.}
\item{freq}{proportion of bungalows}
}
}
\source{
\url{www.seai.ie}
}
\usage{
bungalow_probability
}
\description{
Frequencies of bungalows based on region and housing type (q1) from BER dataset
}
\keyword{datasets}
