% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_factors.R
\docType{data}
\name{capacity_factors}
\alias{capacity_factors}
\title{rooftop solar capacity factors}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{orientation}{roof orientation}
\item{roof1}{roof1 capacity factor}
\item{roof2}{roof2 capacity factor}
}
}
\source{
\url{https://www.seai.ie/data-and-insights/seai-statistics/key-statistics/prices/}
}
\usage{
capacity_factors
}
\description{
unshaded roofs by orientation. roof1 is primary roof, roof2 is secondary roof
}
\keyword{datasets}
