% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{ceg_fun}
\alias{ceg_fun}
\title{ceg_fun}
\usage{
ceg_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
euro/kWh
}
\description{
clean export guarantee price euro/kWh to 2022-2050
}
