% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{ceg_price_inflation_fun}
\alias{ceg_price_inflation_fun}
\title{ceg_price_inflation_fun}
\usage{
ceg_price_inflation_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
ceg inflation expectation in decimal units
}
\description{
feed-in tariff (ceg) price inflation expectations in decimal units. This is used in NPV calculations.
}
