% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{ceg_tax_threshold_fun}
\alias{ceg_tax_threshold_fun}
\title{ceg_tax_threshold_fun}
\usage{
ceg_tax_threshold_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\description{
MSS threshold in euros before tax is liable (tax disregard)
}
