% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_demand.R
\docType{data}
\name{cer_demand}
\alias{cer_demand}
\title{Smart grid CER control group annual demand (2010)}
\format{
A data frame with 155 rows and 2 variables:
\describe{
\item{housecode}{CER household code}
\item{demand}{annual demand in kWh (survey year 2010)}
}
}
\source{
CER (2010)
}
\usage{
cer_demand
}
\description{
Smart grid CER control group annual demand (2010)
}
\keyword{datasets}
