% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_qanda.R
\docType{data}
\name{cer_qanda}
\alias{cer_qanda}
\title{reduced CER survey questions and answers}
\format{
A data frame with 60 rows and 4 variables:
\describe{
\item{question}{survey question}
\item{answercode}{answer code}
\item{answer}{long from of answer}
\item{code}{code}
}
}
\source{
CER 2012
}
\usage{
cer_qanda
}
\description{
useful questions from CER dataset
}
\keyword{datasets}
