% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_questions.R
\docType{data}
\name{cer_questions}
\alias{cer_questions}
\title{questions from reduced CER survey}
\format{
A data frame with 12 rows and 2 variables:
\describe{
\item{code}{code}
\item{question}{question}
}
}
\source{
CER
}
\usage{
cer_questions
}
\description{
questions from reduced CER survey
}
\keyword{datasets}
