% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_survey.R
\docType{data}
\name{cer_survey}
\alias{cer_survey}
\title{CER control group survey data}
\format{
A data frame with 133 rows and 14 variables:
\describe{
\item{housecode}{CER household code}
\item{gender}{gender}
\item{age}{age}
\item{class}{profession}
\item{adults}{calculated total imports in kWh}
\item{children}{calculated total exports in kWh}
\item{housing_type}{calculated total exports in kWh}
\item{housing_ownership}{calculated total exports in kWh}
\item{housing_age}{age range of property}
\item{floor_area}{floor area in m2}
\item{bedrooms}{no of bedrooms}
\item{education}{education level}
\item{demand}{kWh annual demand from CER dataset}
#' }
}
\source{
CER
}
\usage{
cer_survey
}
\description{
from CER 2010 dataset
}
\keyword{datasets}
