% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_systems2.R
\docType{data}
\name{cer_systems2}
\alias{cer_systems2}
\title{Calculated annual grid imports and exports as a function of installed solar and battery capacity for 155 households}
\format{
A data frame with 3,628,395 rows and 7 variables:
\describe{
\item{housecode}{CER household code}
\item{solar1}{installed solar capacity 0-25kW in increments of 0.5kW on first half-roof}
\item{solar2}{installed solar capacity 0-25kW in increments of 0.5kW on second half-roof}
\item{battery}{installed battery capacity 0-100kW in increments of 2.5KWh}
\item{imports}{calculated total imports in kWh}
\item{exports}{calculated total exports in kWh}
\item{aspect}{orientation of house}
\item{demand}{household annual demand (consumption) in kWh}
#' }
}
\source{
CER, pvWatts (NREL)
}
\usage{
cer_systems2
}
\description{
solar PV + battery system import and export table for 155 CER control group households. Used to find the financially optimal system.
}
\details{
derived from CER 2010 dataset
}
\keyword{datasets}
