% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_params_2018_long.R
\docType{data}
\name{cost_params_2018_long}
\alias{cost_params_2018_long}
\title{pricing and cost estimates 2018}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{parameter}{parameter}
\item{value}{value. Costs in euros.}
}
}
\source{
multiple
}
\usage{
cost_params_2018_long
}
\description{
useful questions from CER dataset
}
\keyword{datasets}
