% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{electricity_demand_factor_fun}
\alias{electricity_demand_factor_fun}
\title{electricity_demand_factor_fun}
\usage{
electricity_demand_factor_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
demand factor
}
\description{
mean household demand compared to 2010 (year of CER dataset).
}
