% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{electricity_price_inflation_fun}
\alias{electricity_price_inflation_fun}
\title{electricity_price_inflation_fun}
\usage{
electricity_price_inflation_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
electricity inflation expectation in decimal units
}
\description{
inflation expectations in decimal units. This is used in NPV calculations.
}
