% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_utils.R
\docType{data}
\name{empirical_utils}
\alias{empirical_utils}
\title{empirical utility functions for q9_1 (bills), qsp21 (social) and theta (barrier)}
\format{
A data frame with 7 rows and 3 variables:
\describe{
\item{code}{variable code}
\item{answercode}{survey response code}
\item{du_average}{median partial utility}
}
}
\usage{
empirical_utils
}
\description{
This version assumes pv_data_oo i.e. owner-occupiers of non-apartments only. Generated from pvcalibrater.
}
\keyword{datasets}
