% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{fast_params}
\alias{fast_params}
\title{fast_params}
\usage{
fast_params(params_long)
}
\arguments{
\item{params_long}{long format dataframe with columns "parameter" and "value"}
}
\value{
environment object
}
\description{
helper function to convert a long format dataframe to an environment object, used for fast access to scenario parameters
}
