% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{finance_rate_fun}
\alias{finance_rate_fun}
\title{finance_rate_fun}
\usage{
finance_rate_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
decimal rate (scalar)
}
\description{
historical and future solar PV finance rates. Allows for low interest load subvention.
}
