% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geo_sum_c}
\alias{geo_sum_c}
\title{geo_sum_c}
\usage{
geo_sum_c(d, term)
}
\arguments{
\item{d}{First term}

\item{term}{number of terms}
}
\value{
the sum d+d^2 + .. d^term
}
\description{
The geometric sum d+d^2 + .. d^term
}
