% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_adopters}
\alias{get_adopters}
\title{get_adopters}
\usage{
get_adopters(abm)
}
\arguments{
\item{abm}{output of runABM}
}
\description{
returns the ensemble-averaged fraction of households that adopt during a particular interval.
also includes the fraction of households that augment an existing syste, share of battery adopters,
share of battery augmenters and share that augment their battery only.
}
