% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_available_area_remaining}
\alias{get_available_area_remaining}
\title{get_available_area_remaining}
\usage{
get_available_area_remaining(abm)
}
\arguments{
\item{abm}{output of runABM}
}
\value{
available areas in km2
}
\description{
the remaining available roof area by aspect and date
}
