% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer.R
\name{get_cer_sys_utils}
\alias{get_cer_sys_utils}
\title{get_cer_sys_utils}
\usage{
get_cer_sys_utils(params)
}
\arguments{
\item{params}{parameter set (output of fast_params)}
}
\value{
dataframe with utility column du
}
\description{
financial utility (NPV0-NPV)/NPV0 calculated for each combination of solar and battery capacity
}
