% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_energy}
\alias{get_energy}
\title{get_energy}
\usage{
get_energy(abm, housing_stock = 1.15)
}
\arguments{
\item{abm}{list output of runABM}

\item{housing_stock}{owner-occupier housing stock in millions (default 1.1)}
}
\value{
dataframe with annual values
}
\description{
annual total energetics (demand,solar,imports,exports in GWh) and self-sufficiency metrics
}
