% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_installed}
\alias{get_installed}
\title{get_installed}
\usage{
get_installed(abm, groupvars = NULL)
}
\arguments{
\item{abm}{output of runABM}

\item{groupvars}{additional grouping variables (default NULL)}
}
\description{
median size of solar and battery capacities installed at each time step
}
