% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer.R
\name{get_kWh_from_bills}
\alias{get_kWh_from_bills}
\title{get_kWh_from_bills}
\usage{
get_kWh_from_bills(b, mu = 8.53)
}
\arguments{
\item{b}{reported bill}

\item{mu}{kWh bill adjustment factor (seasonality). No seasonality corresponds to mu=6.}
}
\value{
kWh demand
}
\description{
converts reported bi-monthly electricity bill in 2018 survey to an annual kWh consumption. Ensures that mean demand matches SEAI estimate using default seasonality factor mu=8.52
}
