% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{get_network_characteristics}
\alias{get_network_characteristics}
\title{get_network_characteristics}
\usage{
get_network_characteristics(society, g)
}
\arguments{
\item{society}{society used to construct homophilous network}

\item{g}{homophilous social network}
}
\value{
table
}
\description{
get_network_characteristics
}
