% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{get_roofsection_solar_area}
\alias{get_roofsection_solar_area}
\title{get_roofsection_solar_area}
\usage{
get_roofsection_solar_area(
  house_type,
  q1,
  region,
  demand,
  roof_floor_ratio = 2/sqrt(3),
  usable_roof_fraction = 0.75
)
}
\arguments{
\item{house_type}{housing type is "bungalow" or "two-storey"}

\item{q1}{integer in 2:4 describing housing type (2=terraced, 4=detached, 3=semi-d)}

\item{region}{the region integer in 1:4}

\item{demand}{annual electricity demand in kWh}

\item{roof_floor_ratio}{2/sqrt(3) for 30 degree pitch}

\item{usable_roof_fraction}{fraction of rooftop usable for solar PV default 0.8}
}
\value{
potential half-roof area for solar
}
\description{
generate a half rooftop solar area potential stochastically (log normal distribution). This is calibrated to BER data up to 2023 separately for
bungalows and two-storey dwellings. Bungalows have a higher solar potential because of their relatively larger footprint. This function uses
a detailed model based on house location (region), housing_type (q1) and number of storeys (1 or 2).
}
