% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{get_rooftop_solar_area}
\alias{get_rooftop_solar_area}
\title{get_rooftop_solar_area}
\usage{
get_rooftop_solar_area(
  housing_type,
  demand,
  roof_floor_ratio = 2/sqrt(3),
  usable_roof_fraction = 0.75
)
}
\arguments{
\item{housing_type}{housing type is "bungalow" or "two-storey"}

\item{demand}{annual electricity demand in kWh}

\item{roof_floor_ratio}{2/sqrt(3) for 30 degree pitch}

\item{usable_roof_fraction}{fraction of rooftop usable for solar PV default 0.8}
}
\value{
potential half-roof area for solar
}
\description{
generate a haf rooftop solar capacity area potential stochastically (log normal distribution). This is calibrated to BER data up to 2023 separately for
bungalows and two-storey dwellings. Bungalows have a higher solar potential because of their relatively larger footprint.
}
