% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{get_rooftop_solar_potential}
\alias{get_rooftop_solar_potential}
\title{get_rooftop_solar_potential}
\usage{
get_rooftop_solar_potential(
  housing_type,
  demand,
  roof_floor_ratio = 2/sqrt(3),
  usable_roof_fraction = 0.85,
  kWp_per_m2 = 0.15,
  shading_factor = 0.85
)
}
\arguments{
\item{housing_type}{housing type is "bungalow" or "two-storey"}

\item{demand}{annual electricity demand in kWh}

\item{roof_floor_ratio}{2/sqrt(3) for 30 degree pitch}

\item{usable_roof_fraction}{fraction of rooftop usable for solar PV default 0.8}

\item{kWp_per_m2}{kWp per m2 (technical parameter) default 0.15}

\item{shading_factor}{mean capacity factor reduction factor due to shading}
}
\value{
potential installed solar capacity kWp on half-roof
}
\description{
generate a rooftop solar capacity potential stochastically (log normal distribution). This is calibrated to BER data up to 2023 separately for
bungalows and two-storey dwellings. Bungalows have a higher solar potential because of their relatively larger footprint. A stochastic shading factor is included via a shading_factor parameter.
}
