% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{get_self_consumption}
\alias{get_self_consumption}
\title{get_self_consumption}
\usage{
get_self_consumption(solar_capacity, exports, pv_capacity_factor = 0.094)
}
\arguments{
\item{solar_capacity}{install pv capacity}

\item{exports}{exports to grid}

\item{pv_capacity_factor}{capacity factor default using Birr pv1 time-series 0.094 (Dublin is 0.098)}
}
\value{
metric value
}
\description{
self consumption metric
}
