% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{get_self_sufficiency}
\alias{get_self_sufficiency}
\title{get_self_sufficiency}
\usage{
get_self_sufficiency(imports, demand)
}
\arguments{
\item{imports}{total annual imports kWh}

\item{demand}{total annual demand kWh}
}
\value{
metric value
}
\description{
get_self_sufficiency
}
