% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{get_shaded_sys}
\alias{get_shaded_sys}
\title{get_shaded_sys}
\usage{
get_shaded_sys(cer_sys)
}
\arguments{
\item{cer_sys}{cer_system data (solar1 solar2 battery imports exports) with shading parameters for roof 1 & 2}
}
\value{
cer_system data with imports and exports including effect of shading
}
\description{
estimated effect on annual energy imports and exports computed from a cer_sys dataframe
}
