% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{get_sys_dnpv}
\alias{get_sys_dnpv}
\title{get_sys_dnpv}
\usage{
get_sys_dnpv(
  params,
  demand,
  old_imports,
  old_exports,
  old_solar1,
  old_solar2,
  old_battery,
  new_imports,
  new_exports,
  d_solar1,
  d_solar2,
  d_battery,
  include_grant = T,
  customer = "domestic"
)
}
\arguments{
\item{params}{cost parameter long-format dataframe}

\item{demand}{annual household electrical energy demand (kWh) (from smart meter dataset).}

\item{old_imports}{annual energy (kWh) imports with existing system (old_solar1, old_solar2, old_battery)}

\item{old_exports}{annual energy exports (kWh) with existing system (old_solar1, old_solar2, old_battery)}

\item{old_solar1}{existing installed solar capacity on half roof1}

\item{old_solar2}{existing installed solar capacity on half roof2}

\item{old_battery}{currently installed battery}

\item{new_imports}{annual energy (kWh) import with new (proposed) system (old_solar1 + d_solar1, old_solar2 + d_solar2, old_battery+d_battery)}

\item{new_exports}{annual energy exports (kWh) with new (proposed) system (old_solar1 + d_solar1, old_solar2 + d_solar2, old_battery+d_battery)}

\item{d_solar1}{additional solar capacity on roof1}

\item{d_solar2}{additional solar capacity on roof2}

\item{d_battery}{kWh battery size increment in steps of 2.5 kWh}

\item{include_grant}{True or False}

\item{customer}{"domestic" is only option at the moment}
}
\value{
list consisting of npv with investment, npv with no investment, difference
}
\description{
same as get_sys_npv but returns a scalar only (npv)
}
