% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{get_sys_npv}
\alias{get_sys_npv}
\title{get_sys_npv}
\usage{
get_sys_npv(
  params,
  demand,
  imports,
  exports,
  solar_capacity,
  battery_capacity,
  include_grant = T,
  customer = "domestic"
)
}
\arguments{
\item{params}{cost parameter long-format dataframe}

\item{demand}{annual power demand in kWh e.g. 5000 kWh}

\item{imports}{annual grid imports in kWh}

\item{exports}{annual exports to grid in kWh}

\item{solar_capacity}{installed solar PV capacity in kW}

\item{battery_capacity}{installed battery capacity in kWh}

\item{include_grant}{include SEAI grant T/F}

\item{customer}{"domestic" only at the moment}
}
\value{
dataframe with columns "capital cost", "amortisation","annual_bill","bill_saving","MSS_revenue","npv_loan_period"
}
\description{
returns the npv of power bills and financing on a domestic solar + battery system
}
