% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily.R
\docType{data}
\name{homophily}
\alias{homophily}
\title{Agent homophily}
\format{
A data frame with 8 rows and 2 variables:
\describe{
\item{variable}{demographic variable}
\item{weights}{homophily weight}
}
}
\source{
subjective
}
\usage{
homophily
}
\description{
Demographic variables used to build homophilous network for pv adopters
}
\keyword{datasets}
