% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{is_bungalow}
\alias{is_bungalow}
\title{is_bungalow}
\usage{
is_bungalow(qc1, region)
}
\arguments{
\item{qc1}{qc1=1 rural qc1=2 urban}

\item{region}{Dublin, Leinster, Munster, Connulster}
}
\value{
"bungalow" or "two-storey"
}
\description{
Is this dwelling a bungalow or not based on urban/rural (qc1) and region? Consistent with overall bungalow share of 20\% and lower number of bungalows in urban areas
in BER data.
}
