% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{is_bungalow2}
\alias{is_bungalow2}
\title{is_bungalow2}
\usage{
is_bungalow2(q1, region)
}
\arguments{
\item{q1}{2:5 2=Detached 3=semi-D 4= terraced 5=other. 5 is reassigned to other 2:4.}

\item{region}{Dublin, Leinster, Munster, Connulster}
}
\value{
"bungalow" or "two-storey"
}
\description{
Probabilistic assignment of number of floor (1 or 2) based on housing type (q1) and region.
Based on ~800k BER data sample and consistent with overall bungalow frequency of ~ 21\%
}
