% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{kWp_per_m2_fun}
\alias{kWp_per_m2_fun}
\title{kWp_per_m2_fun}
\usage{
kWp_per_m2_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{decimal time}
}
\value{
kWp per m2
}
\description{
solar panel efficiency (peak power per m2) at time yeartime
}
