% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer.R
\name{map_survey_to_cer}
\alias{map_survey_to_cer}
\title{map_survey_to_cer}
\usage{
map_survey_to_cer(params, lambda = 2)
}
\arguments{
\item{params}{parameter values for mapping (e_demand_factor)}

\item{lambda}{strictness parameter}
}
\description{
mapping of pv survey pwner-occupier data to CER housecodes based on matching annual demand. Higher lambda matches more strictly.
Lower lambda allows more randomness in mapping. pv_survey demand data is re-scaled to match historical mean household demand estimates
Apartments are excluded.
}
