% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pv_cost_fun}
\alias{pv_cost_fun}
\title{pv_cost_fun}
\usage{
pv_cost_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario parameters dataframe}

\item{yeartime}{decimal time}
}
\value{
scalar cost in euros/kWh
}
\description{
solar battery cost in euros/kWh
}
