% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pv_install_cost_fun}
\alias{pv_install_cost_fun}
\title{pv_install_cost_fun}
\usage{
pv_install_cost_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
scalar cost in euros
}
\description{
pv_install_cost_fun
}
