% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_qanda.R
\docType{data}
\name{pv_qanda}
\alias{pv_qanda}
\title{Survey questions and answers related to solar PV}
\format{
A data frame with 126 rows and 4 variables:
\describe{
\item{code}{code label}
\item{question}{long form of question}
\item{answercode}{answercode}
\item{answer}{long form of answer}
}
}
\source{
2018 survey
}
\usage{
pv_qanda
}
\description{
From 2018 survey
}
\keyword{datasets}
