% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_questions.R
\docType{data}
\name{pv_questions}
\alias{pv_questions}
\title{Survey questions related to solar PV}
\format{
A data frame with 27 rows and 2 variables:
\describe{
\item{code}{code label}
\item{question}{long form of question}
}
}
\source{
2018 survey
}
\usage{
pv_questions
}
\description{
From 2018 survey
}
\keyword{datasets}
