% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_survey.R
\docType{data}
\name{pv_survey}
\alias{pv_survey}
\title{Reduced solar PV survey data}
\format{
A data frame with 1208 rows and 27 variables:
\describe{
\item{age}{ 1: 18-24 .... 5: 55+}
\item{gender}{male =1 female =2}
\item{class}{social class}
\item{region}{Dublin, Leinster, Munster, Connulster}
\item{qc1}{Urban or rural}
\item{q1}{Property type}
\item{q3}{Ownership}
\item{q5}{Year of construction}
\item{q7}{Reponsibility for energy decisions}
\item{q9_1}{Bi-monthly bill (average euros)}
\item{q10b}{Perception of electricity use}
\item{q15}{Early or late technology adopter}
\item{q16}{Number of associates who discuss/share technology information (network degree)}
\item{q17b}{Risk aversion Likert scale 1= very risk averse, 5= risk seeking}
\item{q17c}{Willingness to delay gratification Likert scale 1= very unwilling, 5= very willing}
\item{q17a_1}{Climate views}
\item{q17a_2}{Environmental views}
\item{q17a_3}{Preference for energy efficient devices}
\item{q17a_5}{Fuel price expectations}
\item{qsp20}{Satisfaction with exisiting electricity source}
\item{qsp21}{Knowledge of households who have installed solar PV}
\item{qj}{Education level}
\item{qk}{Household type}
\item{qh}{Occupation}
\item{qg}{Main earner}
\item{qf}{Family Status}
\item{qsp22_7}{Stated likelihood to consider installing solar PV}
}
}
\source{
2018 survey
}
\usage{
pv_survey
}
\description{
from 2018 survey.
}
\keyword{datasets}
