% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_0.R
\docType{data}
\name{scenario_0}
\alias{scenario_0}
\title{Basic test parameter scenario for pvmicrosimr}
\format{
A data frame with 43 rows and 6 variables:
\describe{
\item{category}{broad parameter category}
\item{variable}{long form name of variable}
\item{parameter}{parameter name}
\item{value}{parameter value}
\item{unit}{parameter units}
\item{comment}{comment}
}
}
\source{
"~/Policy/AgentBasedModels/solarPV/scenario_parameters.xlxs"
}
\usage{
scenario_0
}
\description{
economic and technical parameters 2020-2050
}
\keyword{datasets}
