% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{scenario_params_df}
\alias{scenario_params_df}
\title{scenario_params_df}
\usage{
scenario_params_df(sD, yeartime)
}
\arguments{
\item{sD}{scenario parameters e.g. scenario_0}

\item{yeartime}{decimal time}
}
\value{
long form dataframe containing parameter names and values
}
\description{
builds the parameter set at yeartime from scenario sD
}
