% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_wem.R
\docType{data}
\name{scenario_wem}
\alias{scenario_wem}
\title{Basic test parameter scenario for pvmicrosimr}
\format{
A data frame with 57 rows and 6 variables:
\describe{
\item{category}{broad parameter category}
\item{variable}{long form name of variable}
\item{parameter}{parameter name}
\item{value}{parameter value}
\item{unit}{parameter units}
\item{comment}{comment}
}
}
\source{
"~/Policy/AgentBasedModels/solarPV/scenario_parameters.xlxs" sheet 1
}
\usage{
scenario_wem
}
\description{
economic and technical parameters 2010-2050. Additional historical data is saved in e.g. seai_elec etc
}
\keyword{datasets}
