% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seai_elec.R
\docType{data}
\name{seai_elec}
\alias{seai_elec}
\title{simple annual Irish electricty household demand and price data from SEAI}
\format{
A data frame with 19 rows and 3 variables:
\describe{
\item{year}{year}
\item{kWh}{mean demand}
\item{price}{mean price c/kWh}
}
}
\source{
\url{https://www.seai.ie/data-and-insights/seai-statistics/key-statistics/prices/} \url{https://www.seai.ie/data-and-insights/seai-statistics/key-statistics/residential/}
}
\usage{
seai_elec
}
\description{
2005-2023
}
\keyword{datasets}
