% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{seai_grant}
\alias{seai_grant}
\title{seai_grant}
\usage{
seai_grant(params, s, b)
}
\arguments{
\item{params}{fast scenario parameters}

\item{s}{solar capacity in kW}

\item{b}{battery capacity in kWh}
}
\value{
grant amount in euros
}
\description{
seai grant for domestic solar installation 2022
}
