% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seai_grant_c}
\alias{seai_grant_c}
\title{seai_grant_c}
\usage{
seai_grant_c(params_vec, s, b)
}
\arguments{
\item{params_vec}{scenario parameter vector}

\item{s}{solar capacity in kW}

\item{b}{battery capacity in kWh}
}
\value{
grant amount in euros
}
\description{
seai grant for domestic solar installation (2022 structure)
fast c++ implementation
}
