% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{seai_grant_fast}
\alias{seai_grant_fast}
\title{seai_grant_fast}
\usage{
seai_grant_fast(params, s, b)
}
\arguments{
\item{params}{fast scenario parameters}

\item{s}{solar capacity in kW}

\item{b}{battery capacity in kWh}
}
\value{
grant amount in euros
}
\description{
The fast version of seai_grant
}
