% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{self_sufficiency_fun}
\alias{self_sufficiency_fun}
\title{self_sufficiency_fun}
\usage{
self_sufficiency_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
reduction factor (between 0 & 1)
}
\description{
self-sufficiency premium vs survey value (< 1)
}
